% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assign_cancer_gene_roles}
\alias{assign_cancer_gene_roles}
\title{Function to classify genes as tumor suppressors/proto-oncogenes,
driver etc. based on multiple lines of evidence.}
\usage{
assign_cancer_gene_roles(
  gox_basic = NULL,
  citation_cutoff_cm1 = 5,
  citation_cutoff_cm2 = 10,
  citation_cutoff_cm3 = 20,
  min_sources_driver = 2
)
}
\arguments{
\item{gox_basic}{output from geneOncoX::get_basic()}

\item{citation_cutoff_cm1}{minimum citation count for genes with CancerMine
support only (VERY STRONG)}

\item{citation_cutoff_cm2}{minimum citation count for genes with CancerMine
support only (STRONG)}

\item{min_sources_driver}{minimum number of sources (NCG, CGC, CancerMine,
TCGA, IntoGen) that must contribute to cancer driver gene status}
}
\value{
data frame with cancer gene annotation status
}
\description{
Function to classify genes as tumor suppressors/proto-oncogenes,
driver etc. based on multiple lines of evidence.
}
\keyword{internal}
