% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_virtual_panel_bed}
\alias{create_virtual_panel_bed}
\title{Function to write virtual panel BED files (CPSR)}
\usage{
create_virtual_panel_bed(
  gwas_bed_fpath = NA,
  gox_gencode = NULL,
  gox_panels = NULL,
  gox_predisposition = NULL,
  super_panel = F,
  build = "grch37",
  dest_dir = NA
)
}
\description{
Function to write virtual panel BED files (CPSR)
}
\keyword{internal}
