% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_predisposition.R
\name{get_predisposition}
\alias{get_predisposition}
\title{Get human cancer predisposition genes}
\usage{
get_predisposition(cache_dir = NA, force_download = FALSE)
}
\arguments{
\item{cache_dir}{Local directory for data download}

\item{force_download}{Logical indicating if local cache should be overwritten}
}
\value{
\strong{metadata} - A data frame with 1 row and 6 columns:
\itemize{
\item \emph{source} - gene annotation source
\item \emph{annotation_data} - type of annotations used
\item \emph{url} - URL of annotation resource
\item \emph{citation} - publication to cite for annotation source
(citation; PMID)
\item \emph{version} - version used
\item \emph{abbreviation} - abbreviation used in column names of records
}

\strong{records} - A data frame with 594 rows and 8 columns:
\itemize{
\item \emph{symbol} - official gene symbol
\item \emph{entrezgene} - Entrez gene identifier
\item \emph{moi} - mechanism of inheritance (AD, AR)
\item \emph{predisp_syndrome_cui} - Concept unique identifiers
(CUI, UMLS) - inherited cancer syndromes
\item \emph{predisp_cancer_cui} - Concept unique identifiers (CUI, UMLS) -
inherited cancer conditions
\item \emph{predisp_source} - Sources supporting predisposition gene
PANEL_APP, CGC, TCGA_PANCAN_2018, OTHER, ACMG_SF
\item \emph{phenotypes} - associated cancer phenotypes
\item \emph{mechanism_of_disease} - mechanism of disease)
}
}
\description{
Downloads and retrieves a pre-processed dataset of human cancer
predisposition genes. Genes deemed relevant for cancer predisposition
have been collected from multiple resources:
\itemize{
\item \href{https://panelapp.genomicsengland.co.uk/}{Genomics England PanelApp}
\item \href{https://pubmed.ncbi.nlm.nih.gov/29625052/}{TCGA's PanCancer study}
\item \href{https://cancer.sanger.ac.uk/census}{Cancer Gene Census}
\item Other/curated
}

The dataset comes as a \code{list} object, with two elements:
\itemize{
\item \code{metadata} - a data frame with metadata regarding annotation
resources used
\item \code{records} - a list with records of cancer predisposing gene,
one record per gene
}
}
\details{
\strong{NOTE}: The dataset also contains genes recommended for reporting of
incidental findings (ACMG_SF).
}
\examples{
\dontrun{
library(geneOncoX)
gene_predisp <- get_predisposition(cache_dir = tempdir())
}

}
