% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sort_bed_regions}
\alias{sort_bed_regions}
\title{Function to sort BED data}
\usage{
sort_bed_regions(unsorted_regions, chr_names = F)
}
\arguments{
\item{unsorted_regions}{data frame with unsorted regions}

\item{chr_names}{logical indicating if chromosome names contain 'chr' or not}
}
\description{
Function to sort BED data
}
\keyword{internal}
