% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{write_bed_file}
\alias{write_bed_file}
\title{Function that writes BED data to file}
\usage{
write_bed_file(bed_data, bed_fname, dest_dir = NA)
}
\arguments{
\item{bed_data}{data frame with BED data}

\item{bed_fname}{file name for BED file}

\item{dest_dir}{directory to write BED file}
}
\value{
file name of bgzipped BED file with BED track data
}
\description{
Function that writes BED data to file
}
\keyword{internal}
