% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assign_cancer_gene_evidence}
\alias{assign_cancer_gene_evidence}
\title{Function to classify genes as tumor suppressors/proto-oncogenes,
driver etc. based on multiple lines of evidence.}
\usage{
assign_cancer_gene_evidence(
  gox_basic = NULL,
  min_citations_tsg = 15,
  min_citations_oncogene = 15
)
}
\arguments{
\item{gox_basic}{output from geneOncoX::get_basic()}

\item{min_citations_tsg}{minimum citations (CancerMine) for tsg support}

\item{min_citations_oncogene}{minimum citations
(CancerMine) for oncogene support}
}
\value{
data frame with cancer gene annotations
}
\description{
Function to classify genes as tumor suppressors/proto-oncogenes,
driver etc. based on multiple lines of evidence.
}
\keyword{internal}
