% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assign_cancer_gene_roles}
\alias{assign_cancer_gene_roles}
\title{Function to classify genes as tumor suppressors/proto-oncogenes,
driver etc. based on multiple lines of evidence.}
\usage{
assign_cancer_gene_roles(
  gox_basic = NULL,
  min_citation_support = 5,
  min_citation_support_cm_only = 20,
  min_sources_driver = 2
)
}
\arguments{
\item{gox_basic}{output from geneOncoX::get_basic()}

\item{min_citation_support}{minimum citation count for support from CancerMine}

\item{min_citation_support_cm_only}{minimum citation count for genes annotated
with cancer gene roles from CancerMine only}

\item{min_sources_driver}{minimum number of sources (NCG, CGC, CancerMine,
TCGA, IntoGen) that must contribute to cancer driver gene status}
}
\value{
data frame with cancer gene annotation status
}
\description{
Function to classify genes as tumor suppressors/proto-oncogenes,
driver etc. based on multiple lines of evidence.
}
\keyword{internal}
