% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_virtual_panel_bed}
\alias{create_virtual_panel_bed}
\title{Function to write virtual panel BED files (CPSR)}
\usage{
create_virtual_panel_bed(
  gwas_bed_fpath = NA,
  gox_gencode = NULL,
  gox_panels = NULL,
  gox_predisposition = NULL,
  super_panel = FALSE,
  build = "grch37",
  dest_dir = NA
)
}
\arguments{
\item{gwas_bed_fpath}{path to assembly-specific GWAS BED track}

\item{gox_gencode}{object returned by \code{geneOncoX::get_gencode()}}

\item{gox_panels}{object returned by \code{geneOncoX::get_panels()}}

\item{gox_predisposition}{object returned by
\code{geneOncoX::get_predisposition()}}

\item{super_panel}{logical indicating if a BED files for a CPSR super panel
are to be generated}

\item{build}{genome assembly build (grch37/grch38)}

\item{dest_dir}{destination directory for BED file}
}
\value{
integer indicating success (0) or failure (-1)
}
\description{
Function to write virtual panel BED files (CPSR)
}
\keyword{internal}
